Refactor : 페이지네이션 및 스케줄 관련 컴포넌트 수정
페이지네이션 코드를 수정하고, 스케줄 관련 컴포넌트 수정 작업을 1차적으로 마쳤습니다.
